Updating: minor cli update

+++ src/cli.dats
@@ -50,10 +50,6 @@ extern
 fun
 do_balance(args: list0(string)): void
 
-extern
-fun
-do_query(args: list0(string)): void
-
 (* ****** ****** *)
 
 implement
@@ -64,12 +60,11 @@ cli_start(lines) = let
   val () = println!(" Commands:                                                                     ")
   val () = println!("    blockchain <from block> <to block>        View current state of blockchain ")  
   val () = println!("    balance <miner>                           Get balance of miner             ")
-  val () = println!("    code <lambda-lisp code>                   Code a smart contract            ")
+  val () = println!("    code <-l/q> <lisp code>                   Code a smart contract            ")
   val () = println!("    define <miner>                            Defines a new miner              ")
   val () = println!("    execute <-l/q> <path/to/filename.txt>     Execute a new smart contract     ")
   val () = println!("    exit                                      Exits the application            ")
   val () = println!("    mine <miner>                              Specified miner mines a new block")
-  val () = println!("    query <query-lisp code>                   Make a DB query                  ")
   val () = println!("    transact <from> <to> <ammount>            Create a new transaction         ")
   val () = println!("                                                                               ")
   val () = println!("                                                                               ")
@@ -114,7 +109,6 @@ cli_do(args) =
       | "define" => do_define(args)
       | "execute" => do_execute(args)
       | "mine" => do_mine(args)
-      | "query" => do_query(args)
       | "transact" => do_transact(args)
       | _ => ()
 
@@ -171,29 +165,47 @@ case+ args of
   )
 
 implement
-do_code(args) = let
-  val f = "lam" + get_time() + ".txt"
-  val out = fileref_open_exn(f, file_mode_a)
-  val () = fprint_string(out, encode_usercode(args))
-  val () = fileref_close(out)
-  val parsed = parse_lisp(f)
-in
-  file_write_contract((f, val2str(interp(parsed)), get_gas(parsed)))
-end
-
-
-implement
-do_query(args) = let
-  val f = "lql" + get_time() + ".txt"
-  val out = fileref_open_exn(f, file_mode_a)
-  val () = fprint_string(out, encode_usercode(args))
-  val () = fileref_close(out)
-  val parsed = qparse_lisp(f)
-  val v = interp(parsed)
-  val () = remove_file(f)
-in
-  file_write_state((lisp2sql(parsed), qval2str(v), get_gas(parsed)))
-end
+do_code(args) = 
+case+ args of 
+| nil0() => (println!("must provide coding language"); ())
+| cons0(t, args) =>
+  case+ args of
+  | nil0() => (println!("must provide code"); ())
+  | cons0 _ =>
+  (
+    if t = "-l"
+    then
+    (
+      let
+        val f = "lam" + get_time() + ".txt"
+        val out = fileref_open_exn(f, file_mode_a)
+        val () = fprint_string(out, encode_usercode(args))
+        val () = fileref_close(out)
+        val parsed = parse_lisp(f)
+      in
+        file_write_contract((f, val2str(interp(parsed)), get_gas(parsed)))
+      end
+    )
+    else
+    (
+      if t = "-q"
+      then
+      (
+        let
+          val f = "lql" + get_time() + ".txt"
+          val out = fileref_open_exn(f, file_mode_a)
+          val () = fprint_string(out, encode_usercode(args))
+          val () = fileref_close(out)
+          val parsed = qparse_lisp(f)
+          val v = interp(parsed)
+          val () = remove_file(f)
+        in
+          file_write_state((lisp2sql(parsed), qval2str(v), get_gas(parsed)))
+        end      
+      )
+      else (println!("Invalid argument", t); ())
+    )
+  )  
 
 
 implement
